automatic SQL migration
Introducing 'automatic' version which uses the DB version and the local
SQL file version to decide between update or rollback dynamically.

The local version in the filesystem is the desired state and the schema
in the DB is the current state. Goose with automatic will attempt to
match the DB version to the desired state in local.

To achieve this, it stores the SQL files persistently into the DB.
Hence if the SQL files are removed, it uses the same SQL entries to
perform goose down.